change:データモデルの修正
